third_party/blink/web_tests/platform/generic/fast/js/toString-stack-overflow-expected.txt

This test checks for a regression against https://bugs.webkit.org/show_bug.cgi?id=3743: Incorrect error message given for certain calls.

The test confirms that the correct exception is thrown in the event of a stack overflow during a call to Array.toString.

It is possible that this may need to be updated if WebKit gets an improvement to its JavaScript stack support. Either through increasing the depth of the recursion, or through some other mechanism.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ary.toString() threw exception RangeError: Maximum call stack size exceeded.
PASS successfullyParsed is true

TEST COMPLETE
